refactor(order): 주문 생성 오케스트레이션을 OrderFacade로 이동
- OrderService.createOrder() 로직을 OrderFacade.ready()로 통합
- OrderCommandMapper 제거 및 Facade 내부에서 직접 변환
- PaymentInfo record 도입으로 도메인 레이어 응답 타입 정리
- place API 엔드포인트 제거 (ready에서 주문 생성까지 일괄 처리)
- 불필요한 OrderCriteria, PaymentCriteria 필드 정리
- 테스트 코드 동기화 (Mock 필드 추가, 삭제된 API 테스트 제거)

Closes #35
