+++ apps/commerce-api/src/main/java/com/loopers/application/payment/PaymentFacade.java
@@ -4,28 +4,18 @@ import com.loopers.application.payment.processor.PaymentProcessContext;
 import com.loopers.domain.order.OrderService;
 import com.loopers.application.payment.processor.PaymentProcessor;
 import com.loopers.domain.order.Order;
-import com.loopers.domain.order.OrderCommand;
-import com.loopers.domain.order.OrderNoIssuer;
 import com.loopers.domain.payment.PayKind;
 import com.loopers.domain.payment.PaymentCommand;
 import com.loopers.domain.payment.PaymentMethod;
+import com.loopers.domain.payment.PaymentInfo;
 import com.loopers.domain.payment.PaymentService;
-import com.loopers.domain.product.Product;
-import com.loopers.domain.product.ProductService;
-import com.loopers.domain.user.User;
-import com.loopers.domain.user.UserService;
-import com.loopers.domain.coupon.CouponService;
 import com.loopers.domain.payment.exception.PaymentException;
 import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -34,10 +24,6 @@ public class PaymentFacade {
     private final OrderService orderService;
     private final Map<String, PaymentProcessor> paymentProcessorMap;
     private final PaymentService paymentService;
-    private final ProductService productService;
-    private final UserService userService;
-    private final CouponService couponService;
-    private final OrderNoIssuer orderNoIssuer;
 
     @Transactional
     public void pay(PaymentCriteria.Pay criteria) {
@@ -50,28 +36,10 @@ public class PaymentFacade {
         PaymentProcessor paymentProcessor = paymentProcessorMap.get(criteria.paymentMethod().toString());
         paymentProcessor.pay(PaymentProcessContext.of(criteria));
     }
-    
-    @Transactional
-    public PaymentService.PaymentReadyResult ready(String orderNo, String orderKey, PaymentCriteria.Ready criteria) {
-        Order order = getOrCreateOrder(orderNo, criteria.userId(), criteria.orderItems(), criteria.couponId());
-        order.validatePay();
 
-        PayKind resolvedPayKind = resolvePayKind(criteria.paymentMethod(), criteria.payKind());
-        PaymentCommand.Ready readyCommand = PaymentCommand.Ready.of(
-                order.getId(),
-                order.getOrderNumber(),
-                order.getUserId(),
-                order.getFinalAmount(),
-                criteria.paymentMethod(),
-                resolvedPayKind
-        );
-
-        return paymentService.ready(readyCommand, orderKey);
-    }
-    
     @Transactional
-    public PaymentService.PaymentSessionResult createPaymentSession(String orderNo, String orderKey, PaymentCriteria.PaymentSession criteria) {
-        Order order = getOrCreateOrder(orderNo, criteria.userId(), criteria.orderItems(), criteria.couponId());
+    public PaymentInfo.SessionResult createPaymentSession(String orderNo, String orderKey, PaymentCriteria.PaymentSession criteria) {
+        Order order = orderService.getOrderByOrderNumber(orderNo);
         order.validatePay();
 
         PayKind resolvedPayKind = resolvePayKind(criteria.paymentMethod(), criteria.payKind());
@@ -97,7 +65,7 @@ public class PaymentFacade {
                 order.getUserId(),
                 criteria.couponId()
         );
-        
+
         return paymentService.createPaymentSession(transactionCommand, orderKey);
     }
 
@@ -105,53 +73,6 @@ public class PaymentFacade {
     public void complete(PaymentCriteria.Complete complete) {
         PaymentCommand.Search search = PaymentCommand.Search.of(complete.transactionKey(), complete.orderNumber());
         paymentService.complete(search);
-    }
-
-    private Order getOrCreateOrder(
-            String orderNo,
-            String userId,
-            List<PaymentCriteria.OrderItem> orderItems,
-            Long couponId
-    ) {
-        Optional<Order> existingOrder = orderService.findOrderByOrderNumber(orderNo);
-        if (existingOrder.isPresent()) {
-            return existingOrder.get();
-        }
-
-        if (orderItems == null || orderItems.isEmpty()) {
-            return orderService.getOrderByOrderNumber(orderNo);
-        }
-
-        User user = userService.findByUserId(userId);
-        List<Long> productIds = orderItems.stream()
-                .map(PaymentCriteria.OrderItem::productId)
-                .toList();
-        Map<Long, Product> productMap = productService.findAllById(productIds).stream()
-                .collect(Collectors.toMap(Product::getId, Function.identity()));
-
-        List<OrderCommand.OrderItem> orderCommandItems = orderItems.stream()
-                .map(item -> {
-                    Product product = productMap.get(item.productId());
-                    if (product == null) {
-                        throw new IllegalArgumentException("상품을 찾을 수 없습니다. productId=" + item.productId());
-                    }
-                    return OrderCommand.OrderItem.builder()
-                            .productId(product.getId())
-                            .price(product.getPrice())
-                            .quantity(item.quantity())
-                            .build();
-                })
-                .toList();
-
-        Long totalAmount = orderCommandItems.stream()
-                .mapToLong(item -> item.getPrice() * item.getQuantity())
-                .sum();
-
-        Long discountAmount = couponService.calculateDiscountAmount(user.getId(), couponId, totalAmount);
-
-        OrderCommand command = OrderCommand.of(user.getId(), orderCommandItems, discountAmount);
-        orderService.place(command, orderNo);
-        return orderService.getOrderByOrderNumber(orderNo);
     }
 
     private PayKind resolvePayKind(PaymentMethod paymentMethod, PayKind payKind) {
