+++ apps/commerce-api/src/main/java/com/loopers/domain/order/OrderService.java
@@ -73,4 +73,5 @@ public class OrderService {
         order.paid();
         orderRepository.save(order);
     }
+
 }
