+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentInfo.java
@@ -1,0 +1,14 @@
+package com.loopers.domain.payment;
+
+public class PaymentInfo {
+
+    public record ReadyResult(Long paymentId, PaymentStatus paymentStatus) {}
+
+    public record SessionResult(
+            String orderNo,
+            String paymentKey,
+            Long amount,
+            String paymentUrl,
+            String pgKind
+    ) {}
+}
